// src/@types/Sequence.interface.ts
export interface IINSDSeq {
	"INSDSeq_accession-version": string;
	INSDSeq_comment: string;
	"INSDSeq_create-date": string;
	INSDSeq_definition: string;
	INSDSeq_division: string;
	"INSDSeq_feature-table": any;
	INSDSeq_keywords: string[];
	INSDSeq_length: string;
	INSDSeq_locus: string;
	INSDSeq_moltype: string;
	INSDSeq_organism: string;
	"INSDSeq_other-seqids": any;
	"INSDSeq_primary-accession": string;
	INSDSeq_project: string;
	INSDSeq_references: any;
	INSDSeq_sequence: string;
	INSDSeq_source: string;
	INSDSeq_strandedness: string;
	INSDSeq_topology: string;
	INSDSeq_taxonomy: string;
	INSDSeq_country: string;
}